# turerp/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Turerp ERP API",
    "description": "Multi-tenant SaaS ERP system API\n\n## Authentication\n\nAll endpoints except `/api/v1/auth/login`, `/api/v1/auth/register`, and `/api/v1/auth/refresh` require JWT Bearer token authentication.\n\n## Rate Limiting\n\nAuthentication endpoints are rate limited to 10 requests per minute per IP address with a burst of 3 requests.\n\n## Localization\n\nAll endpoints support per-request localization via the `Accept-Language` header. Supported languages: `en` (default), `tr`.\n\n## API Versioning\n\n- `/api/v1/` - Current stable API (recommended)\n- `/api/auth/`, `/api/users/` - Legacy routes (deprecated, will be removed in v2)",
    "contact": {
      "name": "Turerp Team",
      "email": "info@turerp.com"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/categories": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get all categories (paginated)",
        "operationId": "get_categories",
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "description": "Page number (1-based, default: 1)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "path",
            "description": "Number of items per page (default: 20, max: 100)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of categories"
          },
          "401": {
            "description": "Not authenticated - missing or invalid JWT token"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get all products (paginated)",
        "operationId": "get_products",
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "description": "Page number (1-based, default: 1)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "path",
            "description": "Number of items per page (default: 20, max: 100)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of products"
          },
          "401": {
            "description": "Not authenticated - missing or invalid JWT token"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/products/{product_id}/variants": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get all variants for a product endpoint (requires authentication)",
        "operationId": "get_variants_by_product",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product variants found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductVariantResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated - missing or invalid JWT token"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create product variant endpoint (requires admin role)",
        "operationId": "create_variant",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductVariant"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Product variant created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductVariantResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Not authenticated - missing or invalid JWT token"
          },
          "403": {
            "description": "Forbidden - admin role required"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/variants/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get product variant by ID endpoint (requires authentication)",
        "operationId": "get_variant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Variant ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product variant found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductVariantResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated - missing or invalid JWT token"
          },
          "404": {
            "description": "Variant not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update product variant endpoint (requires admin role)",
        "operationId": "update_variant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Variant ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductVariant"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Product variant updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductVariantResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Not authenticated - missing or invalid JWT token"
          },
          "403": {
            "description": "Forbidden - admin role required"
          },
          "404": {
            "description": "Variant not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete product variant endpoint (requires admin role)",
        "operationId": "delete_variant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Variant ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product variant deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated - missing or invalid JWT token"
          },
          "403": {
            "description": "Forbidden - admin role required"
          },
          "404": {
            "description": "Variant not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "description": "Simple localized success message payload.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}